Fix linting errors, break methods out into modules
